booking_system_backend/requirements.txt
fastapi
fastmcp
uvicorn
sqlalchemy
pydantic[email]
python-dotenv
pytest
pytest-asyncio
pytest-cov
httpx
psycopg2-binary
gunicorn
requests
